mount: remove Windows stubs
