+++ mount/mount_unix.go
@@ -1,4 +1,4 @@
-// +build !windows
+// +build !darwin,!windows
 
 package mount
 
@@ -10,7 +10,37 @@ import (
 	"golang.org/x/sys/unix"
 )
 
-func recursiveUnmount(target string) error {
+// Mount will mount filesystem according to the specified configuration.
+// Options must be specified like the mount or fstab unix commands:
+// "opt1=val1,opt2=val2". See flags.go for supported option flags.
+func Mount(device, target, mType, options string) error {
+	flag, data := parseOptions(options)
+	return mount(device, target, mType, uintptr(flag), data)
+}
+
+// Unmount lazily unmounts a filesystem on supported platforms, otherwise does
+// a normal unmount. If target is not a mount point, no error is returned.
+func Unmount(target string) error {
+	err := unix.Unmount(target, mntDetach)
+	if err == nil || err == unix.EINVAL {
+		// Ignore "not mounted" error here. Note the same error
+		// can be returned if flags are invalid, so this code
+		// assumes that the flags value is always correct.
+		return nil
+	}
+
+	return &mountError{
+		op:     "umount",
+		target: target,
+		flags:  uintptr(mntDetach),
+		err:    err,
+	}
+}
+
+// RecursiveUnmount unmounts the target and all mounts underneath, starting
+// with the deepest mount first. The argument does not have to be a mount
+// point itself.
+func RecursiveUnmount(target string) error {
 	// Fast path, works if target is a mount point that can be unmounted.
 	// On Linux, mntDetach flag ensures a recursive unmount.  For other
 	// platforms, if there are submounts, we'll get EBUSY (and fall back
@@ -33,7 +63,7 @@ func recursiveUnmount(target string) error {
 
 	var suberr error
 	for i, m := range mounts {
-		err = unmount(m.Mountpoint, mntDetach)
+		err = Unmount(m.Mountpoint)
 		if err != nil {
 			if i == len(mounts)-1 { // last mount
 				return fmt.Errorf("%w (possible cause: %s)", err, suberr)
@@ -48,21 +78,4 @@ func recursiveUnmount(target string) error {
 		}
 	}
 	return nil
-}
-
-func unmount(target string, flags int) error {
-	err := unix.Unmount(target, flags)
-	if err == nil || err == unix.EINVAL {
-		// Ignore "not mounted" error here. Note the same error
-		// can be returned if flags are invalid, so this code
-		// assumes that the flags value is always correct.
-		return nil
-	}
-
-	return &mountError{
-		op:     "umount",
-		target: target,
-		flags:  uintptr(flags),
-		err:    err,
-	}
 }
