+++ mount/mounter_freebsd.go
@@ -1,3 +1,5 @@
+// +build freebsd cgo
+
 package mount
 
 /*
