+++ mount/mounter_unsupported.go
@@ -1,7 +1,7 @@
-// +build !linux,!freebsd freebsd,!cgo
+// +build freebsd,!cgo
 
 package mount
 
 func mount(device, target, mType string, flag uintptr, data string) error {
-	panic("Not implemented")
+	panic("cgo required on freebsd")
 }
