+++ mount/sharedsubtree_linux_test.go
@@ -1,4 +1,4 @@
-// +build linux,go1.13
+// +build linux
 
 package mount
