--- mount/unmount_unsupported.go
@@ -1,11 +1,0 @@
-// +build windows
-
-package mount
-
-func recursiveUnmount(_ string) error {
-	panic("Not implemented")
-}
-
-func unmount(_ string, _ int) error {
-	panic("Not implemented")
-}
